Update README.md
This software is after the paper:
Miguel A. Ferrer, Moises Diaz, Cristina Carmona-Duarte, Réjean Plamondon, "iDeLog: Iterative Dual Spatial and Kinematic Extraction of Sigma-Lognormal Parameter", in IEEE Transactions on Pattern Analysis and Machine Intelligence, ISSN: 0162-8828, vol. 42, no. 1, pp 114-125, 1 Jan 2020, doi: 10.1109/TPAMI.2018.2879312.
It is an alternative to scriptstudio software. Please, go over the paper for more details.
It has been developed as a toolbox in Matlab.
The file pruebaiDeLog.m show an example of how to use the toolbox. Please read the help of iDeLog to know the meaning of each input/output parameter.
The file c-001-01.mat is a test signature.
We also provide the function PruebaReconstruirScriptStudioFile.m which read a file .ana (from scriptstudio) and reconstruct the signature
To install the tollbox I Matlab:
1.	Extract the .p files from LognormalsDecomposition.rar file
2.	Save the .p files in LognormalsDecomposition folder
3.	Add LognormalsDecomposition folder to Matlab path
4.	Run Prueba iDeLog
Any suggestion, please send an email to gpds@gi.ulpgc.es